refactor(core): 木の表現を index_tree へ出し、辿る規則を 1 つにする
ディスクの target_path を木表現へ移す反復の下ごしらえ。挙動は変えていない。

木は今後 indexer（index.bin への読み書きと PATH スキャンの篩）と search（索引本体）の
両方が持つ。どちらか一方へ置くと依存が逆流するので、両者が依存する index_tree.rs へ
出した。移したのは resolve_all / resolve_one / ParentCache / NO_PARENT / CHAIN_CAP と
組み替えの本体で、search 側には索引の並べ方（1 件 32 B の構造体の列）と全件走査の
2 つの最適化（PathCursor・cmp_paths の高速路）が残る。

記憶域の並べ方は 2 通りになるが、辿る規則は 1 つしか書かない。TreeNodes が「i 番の
親・aux・名前をどう取るか」だけを抽象化し、walk_to_root と raw_path_into は両方の
並べ方から同じ実装を通る。単型化されるので索引側の速さは落ちない——実測でむしろ
速くなった（全件の組み直し 逐次 28.0 → 24.9 ms / 並列 7.5 → 4.9 ms）。

挙動不変の根拠（決定的な指標はすべて一致）:
- 索引の常駐 35.44 MiB / blocks 411,413
- ロードの確保 724,061 / PATH スキャンの確保 2,066
- 構築の解放 -312,592
- 原文の再構築は実インデックス 312,691 件すべてで target_path とバイト一致
