+++ snotra-core/CLAUDE.md
@@ -51,6 +51,7 @@
 - `folder.rs` — フォルダ内エントリの列挙とフィルタ/ソート（責務は `//!`）
 - `indexer.rs` — スキャン対象の列挙・重複排除とインデックスキャッシュ（責務は `//!`）
 - `query.rs` — クエリ/履歴キーの正規化と文字ビットマスク（責務は `//!`）
+- `index_tree.rs` — `target_path` をフォルダ木の接頭辞共有で表す、オンディスクと索引が共有する表現（責務は `//!`）。**辿る規則（`walk_to_root` / `raw_path_into`）はここが唯一持ち、記憶域の並べ方だけを `TreeNodes` が抽象化する**——`indexer` の並列 Vec と `search` の構造体の列が同じ実装を通ることが、ディスクと索引で木がずれないことの根拠である
 - `binfmt.rs` — `magic` + `version` 付きバイナリ入出力の共通処理（責務は `//!`）
 - `error.rs` — crate 共通の error 型（責務は `//!`）
 - `window_data.rs` — ウィンドウ位置（`window.bin`）の保存/復元（責務は `//!`）
